// src/DynamicAPI/types/InventorySelisih.ts
export type InventorySelisihLine = {
    ORGANIZATION_CODE: string;
    ORGANIZATION_NAME: string;
    FROM_ORGANIZATION_CODE: string | null;
    ITEM_CODE: string;
    ITEM_DESC: string;
    ORGANIZATION_ID: number;
    INVENTORY_ITEM_ID: number;
    ORGANIZATION_ID_1: number;
    DATE_RECEIVED: string;
    LAST_UPDATE_DATE: string;
    LAST_UPDATED_BY: number;
    CREATION_DATE: string;
    CREATED_BY: number;
    LAST_UPDATE_LOGIN: number;
    PRIMARY_TRANSACTION_QUANTITY: number;
    SUBINVENTORY_CODE: string;
    REVISION: string | null;
    LOCATOR_ID: string | number | null;
    CREATE_TRANSACTION_ID: number;
    UPDATE_TRANSACTION_ID: number;
    LOT_NUMBER: string | null;
    ORIG_DATE_RECEIVED: string;
    COST_GROUP_ID: number;
    CONTAINERIZED_FLAG: number;
    PROJECT_ID: string | number | null;
    TASK_ID: string | number | null;
    ONHAND_QUANTITIES_ID: number;
    ORGANIZATION_TYPE: number;
    OWNING_ORGANIZATION_ID: number;
    OWNING_TP_TYPE: number;
    PLANNING_ORGANIZATION_ID: number;
    PLANNING_TP_TYPE: number;
    TRANSACTION_UOM_CODE: string;
    TRANSACTION_QUANTITY: number;
    SECONDARY_UOM_CODE: string | null;
    SECONDARY_TRANSACTION_QUANTITY: number | null;
    IS_CONSIGNED: number;
    LPN_ID: string | number | null;
    STATUS_ID: string | number | null;
    MCC_CODE: string | null;
};

export type InventorySelisihItem = {
    ITEM_CODE: string;
    ITEM_NUMBER: string;
    ITEM_DESCRIPTION: string;
    INVENTORY_ITEM_ID: number;
    ORGANIZATION_ID: number;
    ORGANIZATION_CODE: string;
    ORGANIZATION_NAME: string;
    SUBINVENTORY_CODE: string;
    LOCATOR_ID: string | number | null;
    LOCATOR: string | null;
    LOCATOR_NAME: string | null;
    QUANTITY: number;
    LINES: InventorySelisihLine[];
    LINE_COUNT: number;
};